Updated work with vault certs initial

+++ deploy.sh
@@ -1,11 +1,24 @@
 #!/bin/bash
 
 cd "$( dirname "${BASH_SOURCE[0]}")"
+
+
+if [ ! -d "keys/" ]; then
+    echo "ERROR: keys/ directory doesn't exist"
+    exit 1
+  # Control will enter here if $DIRECTORY doesn't exist.
+fi
+
 kubectl create namespace trusted-identity
+
+kubectl -n trusted-identity delete secret ti-keys-config
+
 
 kubectl -n trusted-identity delete -f deployment/deployment.yaml
 kubectl -n trusted-identity delete -f deployment/revoker-deployment.yaml
 
+# Check 
+kubectl -n trusted-identity create secret generic ti-keys-config --from-file=keys/
 
 # Create rbac bindings
 kubectl -n trusted-identity create -f deployment/tiRBAC.yaml
